// src/reducer/types.ts
export type PaletteCollectionState = {
  collections: Array<GamePaletteCollection>;
  currentCollectionId: GamePaletteCollection["id"] | null;
};

export type PatternTableState = {
  tables: Array<PatternTable>;
  currentTableId: PatternTable["id"] | null;
};

export type State = {
  systemPalettes: Array<SystemPalette>;
  currentSystemPaletteId: SystemPalette["id"];
  paletteCollections: { [key in GamePaletteType]: PaletteCollectionState };
  patternTables: { [key in PatternTableType]: PatternTableState };
  nametables: Array<Nametable>;
  currentNametableId: Nametable["id"] | null;
};

export type GamePaletteCollectionWithColors = {
  id: GamePaletteCollection["id"];
  label: GamePaletteCollection["label"];
  gamePalettes: Array<GamePaletteWithColors>;
};